Temp state

+++ config/config.go
@@ -21,7 +21,7 @@ type Coinbase struct {
 // Websocket contains the list of channels to subscribe with and what info to store
 type Websocket struct {
 	Channels          []coinbasepro.MessageChannel `json:"channels"`
-	Granularity       int                          `json:"granularity"`
+	Granularity       int64                        `json:"granularity"`
 	RawMessages       bool                         `json:"raw_messages"`
 	StoreTransactions bool                         `json:"store_transactions"`
 }
